Move generated go files to subpackages
