+++ endpoints/api_satisfaction.go
@@ -7,18 +7,23 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package endpoints
 
 import (
 	"net/http"
 )
 
-func WorkloadCourseStudentGet(w http.ResponseWriter, r *http.Request) {
+func SatisfactionCourseCourseGet(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
 
-func WorkloadPost(w http.ResponseWriter, r *http.Request) {
+func SatisfactionPost(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
+}
+
+func SatisfactionStudentStudentGet(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
 	w.WriteHeader(http.StatusOK)
 }
