+++ endpoints/api_user.go
@@ -7,7 +7,7 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package endpoints
 
 import (
 	"net/http"
