+++ models/model_array_of_courses.go
@@ -7,7 +7,7 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
 type ArrayOfCourses struct {
 }
