+++ models/model_array_of_ids.go
@@ -7,11 +7,7 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
-type Workload struct {
-
-	CourseId int64 `json:"course_id,omitempty"`
-
-	Type_ string `json:"type,omitempty"`
+type ArrayOfIds struct {
 }
