+++ models/model_course.go
@@ -7,10 +7,9 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
 type Course struct {
-
 	Id int64 `json:"id,omitempty"`
 
 	Name string `json:"name,omitempty"`
