+++ models/model_student_satisfaction.go
@@ -7,13 +7,12 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
-type StudentWorkload struct {
-
-	WorkloadId int64 `json:"workload_id,omitempty"`
+type StudentSatisfaction struct {
+	CourseId int64 `json:"course_id,omitempty"`
 
 	StudentId int64 `json:"student_id,omitempty"`
 
-	Completion int64 `json:"completion,omitempty"`
+	Satisfaction int64 `json:"satisfaction,omitempty"`
 }
