+++ models/model_student_workload.go
@@ -7,13 +7,12 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
-type StudentSatisfaction struct {
-
-	CourseId int64 `json:"course_id,omitempty"`
+type StudentWorkload struct {
+	WorkloadId int64 `json:"workload_id,omitempty"`
 
 	StudentId int64 `json:"student_id,omitempty"`
 
-	Satisfaction int64 `json:"satisfaction,omitempty"`
+	Completion int64 `json:"completion,omitempty"`
 }
