+++ models/model_user.go
@@ -7,10 +7,9 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
 type User struct {
-
 	Id int64 `json:"id,omitempty"`
 
 	Name string `json:"name"`
