+++ models/model_workload.go
@@ -7,7 +7,10 @@
  * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
  */
 
-package swagger
+package models
 
-type ArrayOfIds struct {
+type Workload struct {
+	CourseId int64 `json:"course_id,omitempty"`
+
+	Type_ string `json:"type,omitempty"`
 }
