--- swagger/logger.go
@@ -1,32 +1,0 @@
-/*
- * Mandatory exercise 1
- *
- * Mandatory exercse 1
- *
- * API version: 1.0.0
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-
-package swagger
-
-import (
-    "log"
-    "net/http"
-    "time"
-)
-
-func Logger(inner http.Handler, name string) http.Handler {
-    return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        start := time.Now()
-
-        inner.ServeHTTP(w, r)
-
-        log.Printf(
-            "%s %s %s %s",
-            r.Method,
-            r.RequestURI,
-            name,
-            time.Since(start),
-        )
-    })
-}
